+++ packages/accounts-controller/jest.config.js
@@ -14,8 +14,6 @@ module.exports = merge(baseConfig, {
   // The display name when running multiple projects
   displayName,
 
-  coveragePathIgnorePatterns: ['./src/tests'],
-
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {